Renames and restructures classes to adhere to java best practices
