+++ FDAndroidClient/src/main/java/com/freedomotic/freedomotic/fragment/ObjectListFragment.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009-2014 Freedomotic team http://freedomotic.com
+ * Copyright (c) 2009-2017 Freedomotic team http://freedomotic.com
  *
  * This file is part of Freedomotic
  *
@@ -17,7 +17,7 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-package com.freedomotic.freedomotic;
+package com.freedomotic.freedomotic.fragment;
 
 import android.app.Activity;
 import android.app.ListFragment;
