+++ FDAndroidClient/src/main/res/layout/activity_object_detail.xml
@@ -22,5 +22,5 @@
     android:id="@+id/object_detail_container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ObjectDetailActivity"
+    tools:context=".activity.ObjectDetailActivity"
     tools:ignore="MergeRootFrame" />
