+++ FDAndroidClient/src/main/res/layout/fragment_object_detail.xml
@@ -25,4 +25,4 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:textIsSelectable="true"
-    tools:context=".ObjectDetailFragment" />
+    tools:context=".fragment.ObjectDetailFragment" />
